+++ app/profile/_layout.tsx
@@ -26,7 +26,6 @@ export default function Layout() {
         options={{
           title: "Profile",
         }}
-        
       />
     </Stack>
   );
