+++ app/projects/details/[projectID].js
@@ -13,9 +13,6 @@ import {
 import axios from "axios";
 
 import { useNavigation, useRoute } from "@react-navigation/native";
-import PTHeader from "@/PTComponents/Header";
-import { Picker } from "@react-native-picker/picker";
-import MultiSelect from "react-native-multiple-select";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { router } from "expo-router";
